requirements.txt
aiogram~=3.0.0rc2
python-dotenv~=1.0.0